+++ CHANGELOG.md
@@ -10,6 +10,8 @@ ChangeLog
 - IMPROVEMENT: disable settings tab when corresponding engine is active
 - IMPROVEMENT: perform a full open/close operation for all three engines
 - IMPROVEMENT: better optical traffic indicators in status bar
+- IMPROVEMENT: use WebSocket connection for regular command events
+- IMPROVEMENT: pass-through render log messages to control client
 - CLEANUP: log the peer information (client mode) in server
 - BUGFIX: do not copy ".map" files in build process for production builds
 - UPDATE: upgrade dependencies